Get latest data from Washington Post Github Repo

Checks if new data is available from the Washington Post Github repo and returns the most current dataset. Returns list with police shootings and agencies dataframes for version 2. Optionally, the user can also select version 1 of the dataset.

Usage

latest_data(version = 2)

Arguments

version

the dataset version number.

Value

tibble

Examples

latest_data()
#> update available...returning latest version
#> $police_shootings
#> # A tibble: 8,004 × 19
#>       id date       threat_…¹ flee_…² armed…³ city  county state latit…⁴ longi…⁵
#>    <dbl> <date>     <chr>     <chr>   <chr>   <chr> <chr>  <chr>   <dbl>   <dbl>
#>  1     3 2015-01-02 point     not     gun     Shel… Mason  WA       47.2  -123. 
#>  2     4 2015-01-02 point     not     gun     Aloha Washi… OR       45.5  -123. 
#>  3     5 2015-01-03 move      not     unarmed Wich… Sedgw… KS       37.7   -97.3
#>  4     8 2015-01-04 point     not     replica San … San F… CA       37.8  -122. 
#>  5     9 2015-01-04 point     not     other   Evans Weld   CO       40.4  -105. 
#>  6    11 2015-01-04 attack    not     gun     Guth… Logan  OK       35.9   -97.4
#>  7    13 2015-01-05 shoot     car     gun     Chan… Maric… AZ       33.3  -112. 
#>  8    15 2015-01-06 point     not     gun     Assa… Saline KS       38.7   -97.6
#>  9    16 2015-01-06 accident  not     unarmed Burl… Des M… IA       40.8   -91.1
#> 10    17 2015-01-06 point     not     replica Knox… Alleg… PA       40.4   -80.0
#> # … with 7,994 more rows, 9 more variables: location_precision <chr>,
#> #   name <chr>, age <dbl>, gender <chr>, race <chr>, race_source <chr>,
#> #   was_mental_illness_related <lgl>, body_camera <lgl>, agency_ids <chr>, and
#> #   abbreviated variable names ¹​threat_type, ²​flee_status, ³​armed_with,
#> #   ⁴​latitude, ⁵​longitude
#> 
#> $agencies
#> # A tibble: 3,104 × 6
#>       id name                                type         state oricodes total…¹
#>    <dbl> <chr>                               <chr>        <chr> <chr>      <dbl>
#>  1  3145 Abbeville County Sheriff's Office   sheriff      SC    SC00100        1
#>  2  2576 Aberdeen Police Department          local_police WA    WA01401        1
#>  3  2114 Abilene Police Department           local_police TX    TX22101        3
#>  4  2088 Abington Township Police Department local_police PA    PA04601        1
#>  5  3187 Acadia Parish Sheriff's Office      sheriff      LA    LA00100        1
#>  6  3375 Acworth Police Department           local_police GA    GA03305        1
#>  7  1241 Ada County Sheriff's Office         sheriff      ID    ID00100        2
#>  8  1615 Adair County Sheriff's Office       sheriff      OK    OK00100        1
#>  9  1978 Adams County Sheriff's Department   sheriff      CO    CO00100       12
#> 10   897 Adams County Sheriff's Department   sheriff      ID    ID00200        1
#> # … with 3,094 more rows, and abbreviated variable name ¹​total_shootings
#>