Washington Post Police Shooting Raw Dataset Version 2

The Washington Post has been compiling a database of every fatal shooting in the United States by a police officer in the line of duty since 2015. The Post created version 2 of this dataset in 2022 to refine and better define data categories, as well as improve documentation about how the data is obtained and classified. In addition to the altering of fields and formatting, a separate agencies dataset has been added which includes information on the law enforcement agencies which have been involved in at least one fatal shooting since 2015. Death record data about each incident and victim. The file can be downloaded at this URL. Victim information is obtained from any available recordings of the incident, news accounts, court records and/or official statements.

Usage

policeshootings_v2

Format

A data frame with 19 variables:

id

double A unique identifier for each fatal police shooting incident.

date

double The date of the fatal shooting.

threat_type

character Actions the victim took leading up to the fatal shooting.

flee_status

character How, if at all, was the victim moving relative to officers leading up to the shooting.

armed_with

character What, if anything, was the victim armed with per federal classifications based on NIBRS, the national incident-level crime reporting system. These categories are roughly reflected in local police data and forms. A NIBRS manual is can be found here (see values for data element Type Weapon/Force involved, on PDF page 104/report page 94).

city

character The municipality where the fatal shooting took place

county

character County where the fatal shooting took plce.

state

character The two-letter postal code abbreviation for the state in which the fatal shooting took place.

latitude

double The latitude location of the shooting expressed as WGS84 coordinates, geocoded from addresses. Please note that the precision and accuracy of incident coordinates varies depending on the precision of the input address which is often only available at the block level.

longitude

double The longitude location of the shooting expressed as WGS84 coordinates, geocoded from addresses.

location_precision

character Indicates the precision level of the input which was geocoded to generate the coordinate data.

name

character The name of the victim.

age

double The age of the victim at the time of the incident.

gender

character The gender of the victim. The Post identifies victims by the gender they identified with if reports indicate that it differs from their biological sex.

race

character The race and ethnicity (where known) of the victim. May contain multiple values to accommodate for multi-racial or several racial and ethnic identifications. Race has been included where news accounts, police reports or other official documents specifically mention a victim’s race or where researchers were able to make a visual determination on racial identification through photos. With the introduction of v2 of the database, the Post has began tracking multiple race and ethnicity designations; prior to 2021, only one race or ethnicity was assigned to most victims.

race_source

character Sourcing methodology for victim race data.

was_mental_illness_related

logical Whether news reports have indicated the victim had a history of mental health issues, expressed suicidal intentions or was experiencing mental distress at the time of the shooting.

body_camera

logical Whether news reports have indicated an officer was wearing a body camera and it may have recorded some portion of the incident.

agency_ids

character List of agency ids associated with the death record.

Source

https://github.com/washingtonpost/data-police-shootings/